# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "This is a sample Laravel API documentation using Swagger",
        "contact": {
            "email": "temuru502@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/types/{id}": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "Get a single type",
                "operationId": "0ab8bd19a1e5819dcf0237308ff76c7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Type not found"
                    }
                }
            }
        },
        "/types/create": {
            "post": {
                "tags": [
                    "Types"
                ],
                "summary": "Create a new type",
                "operationId": "24697bc4d2ac1e03e28420d8dd4a4cf2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/types/update/{id}": {
            "put": {
                "tags": [
                    "Types"
                ],
                "summary": "Update an existing type",
                "operationId": "178b0bb8fb57d8cacb37b43f0894a84f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Type not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/types/delete/{id}": {
            "delete": {
                "tags": [
                    "Types"
                ],
                "summary": "Delete a type",
                "operationId": "b2a232a284f4f26d402971593710667c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Type deleted successfully"
                    },
                    "404": {
                        "description": "Type not found"
                    }
                }
            }
        },
...
        "/unverstet/{id}": {
            "get": {
                "tags": [
                    "Universities"
                ],
                "summary": "Get a single university",
                "operationId": "700c119ab9696e44df1880bddc9e599c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "University details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnverstetResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "University not found"
                    }
                }
            }
        },
        "/unverstet/create": {
            "post": {
                "tags": [
                    "Universities"
                ],
                "summary": "Create a new university",
                "operationId": "d258d51173be6e8683a57869176b4d31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnverstetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "University created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnverstetResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/unverstet/update/{id}": {
            "put": {
                "tags": [
                    "Universities"
                ],
                "summary": "Update an existing university",
                "operationId": "191a53bd0e9ffea89d79ff25b9e26071",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnverstetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "University updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnverstetResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "University not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/unverstet/delete/{id}": {
            "delete": {
                "tags": [
                    "Universities"
                ],
                "summary": "Delete a university",
                "operationId": "db3006d30f4ffc22def72a46d09e9aba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "University deleted successfully"
                    },
                    "404": {
                        "description": "University not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TypeRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Example Type"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the type"
                    }
                },
                "type": "object"
            },
            "UnverstetRequest": {
                "title": "University Request",
                "required": [
                    "name",
                    "location"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Sample University"
                    },
                    "location": {
                        "type": "string",
                        "example": "City, Country"
                    }
                },
                "type": "object"
            },
...
            "TypeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Example Type"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the type"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-24T12:34:56"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-24T12:34:56"
                    }
                },
                "type": "object"
            },
            "UnverstetResource": {
                "title": "University Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sample University"
                    },
                    "location": {
                        "type": "string",
                        "example": "City, Country"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}